+++ BitmovinConvivaAnalytics/Classes/ConvivaAnalytics.swift
@@ -1,3 +1,4 @@
+// swiftlint:disable file_length
 //
 //  ConvivaAnalytics.swift
 //  BitmovinConvivaAnalytics
@@ -17,16 +18,12 @@ public final class ConvivaAnalytics: NSObject {
     // MARK: - Conviva related attributes
     let customerKey: String
     let config: ConvivaConfiguration
-    var client: CISClientProtocol
-    var playerStateManager: CISPlayerStateManagerProtocol!
-    var sessionKey: Int32 = NO_SESSION_KEY
+    let analytics: CISAnalytics
+    let videoAnalytics: CISVideoAnalytics
     let contentMetadataBuilder: ContentMetadataBuilder
-    var isSessionActive: Bool {
-        return sessionKey != NO_SESSION_KEY
-    }
+    var isSessionActive: Bool = false
+    var isBumper: Bool = false
 
-    // The BitmovinPlayerListener is used to prevent listener methods to be public and therefore
-    // preventing calling from outside.
     var listener: BitmovinPlayerListener?
 
     // MARK: - Helper
@@ -68,8 +65,6 @@ public final class ConvivaAnalytics: NSObject {
         - player: Bitmovin Player instance to track
         - customerKey: Conviva customerKey
         - config: ConvivaConfiguration object (see ConvivaConfiguration for more information)
-
-     - Throws: Convivas `CISClientProtocol` and `CISClientSettingsProtocol` if an error occurs
      */
     public init?(player: Player,
                  customerKey: String,
@@ -79,24 +74,18 @@ public final class ConvivaAnalytics: NSObject {
         self.customerKey = customerKey
         self.config = config
 
-        let systemInterFactory: CISSystemInterfaceProtocol = IOSSystemInterfaceFactory.initializeWithSystemInterface()
-        let setting: CISSystemSettings = CISSystemSettings()
-
+        if let gatewayUrl = config.gatewayUrl {
+            var settings = [String: Any]()
+            settings[CIS_SSDK_SETTINGS_GATEWAY_URL] = gatewayUrl.absoluteString
+            settings[CIS_SSDK_SETTINGS_LOG_LEVEL] = config.convivaLogLevel.rawValue
+            analytics = CISAnalyticsCreator.create(withCustomerKey: customerKey, settings: settings)
+        } else {
+            analytics = CISAnalyticsCreator.create(withCustomerKey: customerKey)
+        }
         logger = Logger(loggingEnabled: config.debugLoggingEnabled)
         self.contentMetadataBuilder = ContentMetadataBuilder(logger: logger)
 
-        if config.debugLoggingEnabled {
-            setting.logLevel = LogLevel.LOGLEVEL_DEBUG
-        }
-
-        let systemFactory = CISSystemFactoryCreator.create(withCISSystemInterface: systemInterFactory, setting: setting)
-        let clientSetting: CISClientSettingProtocol = try CISClientSettingCreator.create(withCustomerKey: customerKey)
-        if let gatewayUrl = config.gatewayUrl {
-            clientSetting.setGatewayUrl(gatewayUrl.absoluteString)
-        }
-
-        self.client = try CISClientCreator.create(withClientSettings: clientSetting, factory: systemFactory)
-
+        videoAnalytics = analytics.createVideoAnalytics()
         super.init()
 
         listener = BitmovinPlayerListener(player: player)
@@ -117,7 +106,7 @@ public final class ConvivaAnalytics: NSObject {
         - attributes: A dictionary with custom event attributes
      */
     public func sendCustomApplicationEvent(name: String, attributes: [String: String] = [:]) {
-        client.sendCustomEvent(NO_SESSION_KEY, eventname: name, withAttributes: attributes)
+        analytics.reportAppEvent(name, details: attributes)
     }
 
     /**
@@ -129,24 +118,20 @@ public final class ConvivaAnalytics: NSObject {
         - attributes: A dictionary with custom event attributes
      */
     public func sendCustomPlaybackEvent(name: String, attributes: [String: String] = [:]) {
-        if !isSessionActive {
-            logger.debugLog(message: "Cannot send playback event, no active monitoring session")
-            return
-        }
-        client.sendCustomEvent(sessionKey, eventname: name, withAttributes: attributes)
+        analytics.reportAppEvent(name, details: attributes)
     }
 
     // MARK: - external session handling
 
     /**
-     Will update the contentMetadata which are tracked with conviva.
+     Will update the contentMetadata that are tracked with conviva.
 
      If there is an active session only permitted values will be updated and propagated immediately.
      If there is no active session the values will be set on session creation.
 
-     Attributes set via this method will override automatic tracked once.
+     Attributes set via this method will override automatically tracked attributes.
      - Parameters:
-        - metadataOverrides: Metadata attributes which will be used to track to conviva.
+        - metadataOverrides: Metadata attributes which will be used to track for Conviva.
                              @see ContentMetadataBuilder for more information about permitted attributes
      */
     public func updateContentMetadata(metadataOverrides: MetadataOverrides) {
@@ -168,8 +153,8 @@ public final class ConvivaAnalytics: NSObject {
 
      Warning: The integration can only be validated without external session managing. So when using this method we can
      no longer ensure that the session is managed at the correct time. Additional: Since some metadata attributes
-     relies on the players source we can't ensure that all metadata attributes are present at session creation.
-     Therefore it could be that there will be a 'ContentMetadata created late' issue after conviva validation.
+     rely on the player's source we can't ensure that all metadata attributes are present at session creation.
+     Therefore it is possible that we receive a 'ContentMetadata created late' issue after conviva validation.
 
      If no source was loaded (or the itemTitle is missing) and no assetName was set via updateContentMetadata
      this method will throw an error.
@@ -205,6 +190,10 @@ public final class ConvivaAnalytics: NSObject {
         internalEndSession()
     }
 
+    public func release() {
+        videoAnalytics.cleanup()
+        analytics.cleanup()
+    }
     /**
      Sends a custom deficiency event during playback to Conviva's Player Insight. If no session is active it will NOT
      create one.
@@ -221,62 +210,64 @@ public final class ConvivaAnalytics: NSObject {
             return
         }
 
-        client.reportError(sessionKey, errorMessage: message, errorSeverity: severity)
+        videoAnalytics.reportPlaybackError(message, errorSeverity: severity)
         if endSession {
             internalEndSession()
         }
     }
 
     /**
-     Puts the session state in a notMonitored state.
+     Puts the session in a notMonitored state.
      */
-    public func pauseTracking() {
-        // AdStart is the preferred way to pause tracking according to conviva.
-        client.adStart(sessionKey,
-                       adStream: .ADSTREAM_SEPARATE,
-                       adPlayer: .ADPLAYER_SEPARATE,
-                       adPosition: .ADPOSITION_PREROLL)
-        client.detachPlayer(sessionKey)
+    public func pauseTracking(isBumper: Bool) {
+        self.isBumper = isBumper
+        let event: String = self.isBumper ?
+        CISConstants.getEventsStringValue(Events.BUMPER_VIDEO_STARTED) :
+        CISConstants.getEventsStringValue(Events.USER_WAIT_STARTED)
+        videoAnalytics.reportPlaybackEvent(event)
         logger.debugLog(message: "Tracking paused.")
     }
 
     /**
-     Puts the session state from a notMonitored state into the last one tracked.
+     Changes the session state from notMonitored state into the previous state.
      */
     public func resumeTracking() {
-        client.attachPlayer(sessionKey, playerStateManager: playerStateManager)
-        // AdEnd is the preferred way to resume tracking according to conviva.
-        client.adEnd(sessionKey)
+        let event: String = self.isBumper ?
+        CISConstants.getEventsStringValue(Events.BUMPER_VIDEO_ENDED) :
+        CISConstants.getEventsStringValue(Events.USER_WAIT_ENDED)
+        videoAnalytics.reportPlaybackEvent(event)
+        self.isBumper = false
         logger.debugLog(message: "Tracking resumed.")
     }
 
     // MARK: - session handling
     private func setupPlayerStateManager() {
-        playerStateManager = client.getPlayerStateManager()
-        playerStateManager.setPlayerState!(PlayerState.CONVIVA_STOPPED)
-        playerStateManager.setPlayerType!("Bitmovin Player iOS")
-
-        playerStateManager.setCISIClientMeasureInterface?(self)
-
+        videoAnalytics.reportPlaybackMetric(
+            CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE,
+            value: PlayerState.CONVIVA_STOPPED.rawValue)
+        var playerInfo = [String: Any]()
+        playerInfo[CIS_SSDK_PLAYER_FRAMEWORK_NAME] = "Bitmovin Player iOS"
         if let bitmovinPlayerVersion = playerHelper.version {
-            playerStateManager.setPlayerVersion!(bitmovinPlayerVersion)
+            playerInfo[CIS_SSDK_PLAYER_FRAMEWORK_VERSION] = bitmovinPlayerVersion
         }
+        videoAnalytics.setPlayerInfo(playerInfo)
     }
 
     private func internalInitializeSession() {
         buildContentMetadata()
 
-        sessionKey = client.createSession(with: contentMetadataBuilder.build())
-        if !isSessionActive {
-            logger.debugLog(message: "Something went wrong, could not obtain session key")
+        if isSessionActive {
+            logger.debugLog(message: "Returning, session already up")
             return
         }
 
         setupPlayerStateManager()
         updateSession()
 
-        client.attachPlayer(sessionKey, playerStateManager: playerStateManager)
+        videoAnalytics.reportPlaybackRequested(contentMetadataBuilder.build())
+        logger.debugLog(message: "Creating session \(sessionKey) with metadata: \(contentMetadataBuilder)")
         logger.debugLog(message: "Session started")
+        isSessionActive = true
     }
 
     private func updateSession() {
@@ -288,25 +279,34 @@ public final class ConvivaAnalytics: NSObject {
 
         if let videoQuality = player.videoQuality {
             let bitrate = Int(videoQuality.bitrate) / 1000 // in kbps
-            playerStateManager.setBitrateKbps!(bitrate)
-            playerStateManager.setVideoResolutionWidth!(videoQuality.width)
-            playerStateManager.setVideoResolutionHeight!(videoQuality.height)
+            let value = NSValue(cgSize: CGSize(
+                width: CGFloat(videoQuality.width),
+                height: CGFloat(videoQuality.height
+                )))
+            videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_BITRATE, value: bitrate)
+            videoAnalytics.reportPlaybackMetric(
+                CIS_SSDK_PLAYBACK_METRIC_RESOLUTION,
+                value: value.cgSizeValue)
+            videoAnalytics.reportPlaybackMetric(
+                CIS_SSDK_PLAYBACK_METRIC_RENDERED_FRAMERATE,
+                value: player.currentVideoFrameRate)
         }
-
-        client.updateContentMetadata(sessionKey, metadata: contentMetadataBuilder.build())
+        
+        videoAnalytics.setContentInfo(contentMetadataBuilder.build())
+        logger.debugLog(message: "Updating session \(sessionKey) with metadata: \(contentMetadataBuilder)")
     }
 
     private func internalEndSession() {
+        contentMetadataBuilder.reset()
         if !isSessionActive {
             return
         }
 
-        client.detachPlayer(sessionKey)
-        client.cleanupSession(sessionKey)
-        playerStateManager.reset?()
-        client.releasePlayerStateManager(playerStateManager)
-        sessionKey = NO_SESSION_KEY
-        contentMetadataBuilder.reset()
+        videoAnalytics.reportPlaybackEnded()
+
+        isSessionActive = false
+        logger.debugLog(message: "Ending session \(sessionKey)")
+
         playbackStarted = false
         logger.debugLog(message: "Session ended")
     }
@@ -346,15 +346,15 @@ public final class ConvivaAnalytics: NSObject {
         if isStalled && playerState != .CONVIVA_BUFFERING {
             return
         }
-        // do not report any stalling when isStalled false (StallEnded trigered immediatelly after StallStarted)
+        // do not report any stalling when isStalled false (StallEnded triggered immediatelly after StallStarted)
         else if !isStalled && playerState == .CONVIVA_BUFFERING {
             self.logger.debugLog(
                 message: "[ ConvivaAnalytics ] false stalling, not registering to Conviva"
             )
-            return;
+            return
         }
 
-        playerStateManager.setPlayerState!(playerState)
+        videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_PLAYER_STATE, value: playerState.rawValue)
         logger.debugLog(message: "Player state changed: \(playerState.rawValue)")
     }
 }
@@ -386,8 +386,8 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
              internalInitializeSession()
          }
 
-         let message = "\(errorCode) \(errorMessage)"
-         reportPlaybackDeficiency(message: message, severity: .ERROR_FATAL)
+        let message = "\(errorCode) \(errorMessage)"
+        reportPlaybackDeficiency(message: message, severity: .ERROR_FATAL)
     }
 
     func onMuted(_ event: MutedEvent) {
@@ -427,8 +427,8 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
             self.logger.debugLog(
                 message: "[ ConvivaAnalytics ] calling StallStarted after 0.10 seconds"
             )
-             self.onPlaybackStateChanged(playerState: .CONVIVA_BUFFERING)
-             }
+            self.onPlaybackStateChanged(playerState: .CONVIVA_BUFFERING)
+        }
     }
 
     func onStallEnded() {
@@ -450,7 +450,7 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
             return
         }
 
-        playerStateManager.setSeekStart!(Int64(event.to.time * 1000))
+        videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_SEEK_STARTED, value: Int64(event.to.time * 1000))
     }
 
     func onSeeked() {
@@ -459,7 +459,8 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
             return
         }
 
-        playerStateManager.setSeekEnd!(Int64(player.currentTime * 1000))
+        videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_SEEK_ENDED,
+                                            value: Int64(player.currentTime * 1000))
     }
 
     func onTimeShift(_ event: TimeShiftEvent) {
@@ -469,7 +470,7 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
         }
 
         // According to conviva it is valid to pass -1 for seeking in live streams
-        playerStateManager.setSeekStart!(-1)
+        videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_SEEK_STARTED, value: Int64(-1))
     }
 
     func onTimeShifted() {
@@ -477,41 +478,39 @@ extension ConvivaAnalytics: BitmovinPlayerListenerDelegate {
             // See comment in onSeek
             return
         }
-
-        playerStateManager.setSeekEnd!(-1)
+        videoAnalytics.reportPlaybackMetric(CIS_SSDK_PLAYBACK_METRIC_SEEK_ENDED, value: Int64(-1))
     }
 
     #if !os(tvOS)
     // MARK: - Ad events
     func onAdStarted(_ event: AdStartedEvent) {
         let adPosition: AdPosition = AdEventUtil.parseAdPosition(event: event, contentDuration: player.duration)
-        client.adStart(sessionKey, adStream: .ADSTREAM_SEPARATE, adPlayer: .ADPLAYER_CONTENT, adPosition: adPosition)
+        var adAttributes = [String: Any]()
+        adAttributes["c3.ad.position"] = adPosition.rawValue
+        videoAnalytics.reportAdBreakStarted(AdPlayer.ADPLAYER_CONTENT,
+                                            adType: AdTechnology.CLIENT_SIDE, adBreakInfo: adAttributes)
     }
 
     func onAdFinished() {
-        client.adEnd(sessionKey)
+        videoAnalytics.reportAdBreakEnded()
     }
 
     func onAdSkipped(_ event: AdSkippedEvent) {
         customEvent(event: event)
-        client.adEnd(sessionKey)
+        videoAnalytics.reportAdBreakEnded()
     }
 
     func onAdError(_ event: AdErrorEvent) {
         customEvent(event: event)
-        client.adEnd(sessionKey)
+        videoAnalytics.reportAdBreakEnded()
     }
 
     func onAdBreakStarted(_ event: AdBreakStartedEvent) {
         customEvent(event: event)
-        client.detachPlayer(sessionKey)
     }
 
     func onAdBreakFinished(_ event: AdBreakFinishedEvent) {
         customEvent(event: event)
-        if !client.isPlayerAttached(sessionKey) {
-            client.attachPlayer(sessionKey, playerStateManager: playerStateManager)
-        }
     }
     #endif
 
@@ -528,11 +527,5 @@ extension ConvivaAnalytics: UserInterfaceListener {
 
     public func onFullscreenExit(_ event: FullscreenExitEvent) {
         customEvent(event: event)
-    }
-}
-
-extension ConvivaAnalytics: CISIClientMeasureInterface {
-    public func getAverageFrames() -> Int {
-        return Int(player.currentVideoFrameRate)
     }
 }