+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [3.0.0]
+
+- Utilises Conviva's Simple SDK instead of the old interface
+- Added `release()` method for use at end of app lifecycle
+- Internal changes to ConvivaAnalytics and related classes 
+
+### Fixed
+
+- Content length not set correctly (when using `initializeSession()`)
+
 ## [2.0.1]
 
 ### Fixed